Split out parts of app a bit
