+++ .prettierrc
@@ -1,0 +1,2 @@
+trailingComma: "es5"
+tabWidth: 2
